Added Fabric and Twitter Login
Fabric is a twitter sdks that can be downloaded onto android studio. It
has a lot of features and the main one we will be using is the
TwitterKit. It allows us to communicate with the API. Currently you can
login to Twitter and give Glance permissions but nothing more happens.