Added TODOs and Toasts
Added things to do in MainActivity that deal with the Twitter API and
login. Added toasts for when the user does not give Glance permission
or something goes wrong. Added another toasts for when they do give
permission. This is just added for debugging purposes so we can see the
success.