Updates to Login and Support for Tablets
I switched us to a grid view and we now have support for tablets. The
grid view also fixed a few problems we were having with the cards. I
also added login for Instagram and Twitter in the settings. And lastly,
I added an “add accounts later” button on the initial login page.