Fixed Login/Settings Bug
Originally when you logged in on the settings page and then you clicked
the back button it would send you to the login page or callback page,
which the user should never see, and then when you hit it again it
would send you back to settings. I fixed it to send you straight to
main, like it should. I also fixed the flow of the app. When the user
is on the main page (where the posts are) and hits the back button the
app will now close instead of sending you to the login page. I also
added LOG_TAGs for all of the java files.