Updated User Experience
The user will now only see the intro pages when it is their first time
to the app. On the last intro page there is now a button to continue
due to a bug with it going to settings before you even interacted with
the page. The Action Bar has been updated. There can only be one card
flipped over at a time. If another card is flipped over then it will
flip the old one to the front and then flip the new one to the back.
There is a splash screen when you first open the app to help with
loading experience. Will add it to the Settings Page in the next
commit. Updated the logo for the app. Code cleanup.