Updated Info Buttons
The open in twitter and open in Instagram will now open the specific
social media account to my profile. When we hook up the back end then
we will actually be able to send them to the correct user or post. The
retweet and favorite buttons still do not work because we are not
hooked up to the backend yet. The link to the API reference is included
in the click handlers for this function.