Fixing Instagram Feed
Need to implement a way to quickly load posts to the feed. The
preferred way would be to get a page and as we get valid posts upload
them to the server and to the feed. Then when we get done with that
page of JSON move to the next one. Currently it loads in ten pages and
then shows you the contents. Still a little buggy.