Updates to Account UI and Code Clean Up
Cleaned up some more of the code and got rid of some of the useless
files that I found. Updated the UI of the account page so you can
switch between the best friend list and the account info. The account
info view currently has nothing on it besides some filler text for
testing. The posts on the main page now require a long click and no
longer flip. This is to get ready for the new UI updates that we will
be applying on the main page. Focusing on separating content from
interactions.